feat 更新歌曲被添加的状态
