+++ app/renderer/src/App.tsx
@@ -1,4 +1,4 @@
-import React, { useState, createContext } from 'react';
+import React, { useState } from 'react';
 import Header from './components/TitleBar';
 import Menu from './components/Menu';
 import Player from './components/Player';
@@ -6,8 +6,7 @@ import Playing from './pages/Playing';
 import Setting from './pages/Setting';
 import PlayList from './pages/PlayList';
 import styled from 'styled-components';
-import { HowlPlayer } from './utils';
-import { Track, getFile } from './api';
+import PlayContext, { HorenContext } from './components/PlayContext';
 
 const APP = styled.div`
   min-width: 800px;
@@ -38,85 +37,13 @@ const Page = styled.div`
   height: calc(100vh - 136px);
 `;
 
-interface IHorenContext {
-  player: {
-    add: (track: Track) => void;
-    remove: (track: Track) => void;
-    play: (track: Track) => void;
-    next: () => void;
-    prev: () => void;
-    trackList: Track[];
-    currentTrack: Track | null;
-  };
-}
-
-const player = new HowlPlayer<Track>();
-
-export const HorenContext = createContext<IHorenContext>({
-  player: {
-    add: () => {},
-    remove: () => {},
-    play: () => {},
-    next: () => {},
-    prev: () => {},
-    trackList: [],
-    currentTrack: null,
-  },
-});
-
 export type PageName = 'playing' | 'setting';
 
 export default function App() {
   const [page, setPage] = useState<PageName | string>('playing');
-  const [currentTrack, setCurrentTrack] = useState<Track | null>(null);
-  const [trackList, setTrackList] = useState<Track[]>([]);
-
-  const play = (track: Track) => {
-    console.log(trackList);
-    setCurrentTrack(track);
-
-    getFile(track.src).then((res) => {
-      player.currentTrack = { ...track, src: res };
-    });
-
-    if (!includes(trackList, track)) {
-      setTrackList((prev) => [...prev, track]);
-    }
-  };
-
-  const add = (track: Track) => {
-    if (!includes(trackList, track)) setTrackList((prev) => [...prev, track]);
-  };
-
-  const remove = (track: Track) => {
-    setTrackList((prev) => prev.filter((t) => t.title !== track.title));
-  };
-
-  const next = () => {
-    if (currentTrack) {
-      const idx = trackList.indexOf(currentTrack);
-      const length = trackList.length;
-      if (idx < length - 1) {
-        play(trackList[idx + 1]);
-      }
-    }
-  };
-
-  const prev = () => {
-    if (currentTrack) {
-      const idx = trackList.indexOf(currentTrack);
-      if (idx > 0) {
-        play(trackList[idx - 1]);
-      }
-    }
-  };
 
   return (
-    <HorenContext.Provider
-      value={{
-        player: { add, play, remove, prev, next, currentTrack, trackList },
-      }}
-    >
+    <PlayContext>
       <APP>
         <Top className="app-top electron-drag">
           <Header />
@@ -133,12 +60,8 @@ export default function App() {
           <Player />
         </Bottom>
       </APP>
-    </HorenContext.Provider>
+    </PlayContext>
   );
 }
 
-const includes = (tracks: Track[], track: Track) => {
-  for (const t of tracks) {
-    if (t.title === track.title) return true;
-  }
-};
+export { HorenContext };
